Arel::Nodes::ArraySubselect
