+++ lib/arel/extensions/nodes.rb
@@ -4,10 +4,6 @@
 
 module Arel
   module Nodes
-    # postgres only: https://www.postgresql.org/docs/9.1/functions-comparisons.html
-    class ArraySubselect < Arel::Nodes::Unary
-    end
-
     # postgres only: https://www.postgresql.org/docs/9.1/sql-expressions.html
     class TypeCast < Arel::Nodes::Node
       attr_reader :arg
